Responsive Design - viewport reset
add in "maximum-scale=1.2" in the viewport setting, because it seems
that in small screen device, scaling more than 1.2 will result in
overlap of the footer and the nav bar. The down side of setting a
maximum-scale is that some users with visual impair may require larger
characters. I assume, though, that they will not view web pages on a
phone with small screen.